pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hac-mcp"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for scraping Home Access Center (HAC)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "beautifulsoup4",
    "mcp[cli]>=1.0.0",
    "pydantic",
    "python-dotenv"
]

[project.scripts]
hac-mcp = "mcp_server:mcp.main"

[tool.setuptools]
py-modules = ["mcp_server"]
packages = ["hac_client"]
